Project building successfully after complete refactor
