+++ PedalBoard/Pedal/Knobs/Knob.swift
@@ -1,0 +1,12 @@
+//
+//  Knob.swift
+//  PedalCore
+//
+//  Created by Lucas Migge on 02/10/23.
+//
+
+import Foundation
+
+struct Knob {
+    // Namespace umbrella struct for Knob related code
+}
