+++ PedalBoard/Pedal/Models/PedalModel.swift
@@ -12,7 +12,7 @@ extension Pedal {
         var id: String = UUID().uuidString
         var name: String
         var brand: String
-        var knobs: [Knob]
+        var knobs: [Knob.Model]
         
         var signature: String {
             let knobNames = knobs.map { $0.name }
