+++ PedalCore/Providers/LocalDataProvider.swift
@@ -7,24 +7,24 @@
 
 import Foundation
 
-class LocalDataProvider<T: Hashable>: DataProviderProtocol {
+public class LocalDataProvider<T: Hashable>: DataProviderProtocol {
     typealias T = T
     
     private let persistence: any PersistenceProtocol<T>
     
     private(set) var data: [T]
     
-    init(persistence: any PersistenceProtocol<T>) {
+    public init(persistence: any PersistenceProtocol<T>) {
         self.persistence = persistence
         self.data = []
     }
     
-    func update(_ newData: [T]) throws {
+    public func update(_ newData: [T]) throws {
         self.data = newData
         try persistence.save(data)
     }
     
-    func load(_ onLoad: @escaping ([T]) -> Void) throws {
+    public func load(_ onLoad: @escaping ([T]) -> Void) throws {
         try persistence.load { loadedData in
             self.data = loadedData
             onLoad(loadedData)
